ai_backend/requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
flask==2.3.3
flask-cors==4.0.0
pillow==10.0.1
requests==2.31.0 